requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloud-sql-python-connector==1.18.1
contourpy==1.3.0
cryptography==44.0.2
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.57.0
frozenlist==1.6.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.1
gunicorn==21.2.0
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.5.3
MarkupSafe==3.0.2
matplotlib==3.9.4
mlxtend==0.23.4
multidict==6.4.3
narwhals==1.36.0
numpy==2.0.2
packaging==25.0
pandas==2.2.3
pg8000==1.31.2
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.2.1
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.13.1
scramp==1.4.5
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.0
zipp==3.21.0
